requirements.txt
pillow==11.3.*
requests==2.32.*
